+++ src/index.js
@@ -62,6 +62,7 @@ createCommand('prepublish', 'Prepare for a release: validate versions, copy READ
 createCommand('publish', 'Publish updated sub-packages')
 .option('--no-master', 'allow publishing from a non-master branch')
 .option('--no-confirm', 'do not ask for confirmation before publishing')
+.option('--version <new-version>', 'use this version for publishing, instead of asking')
 .option('--publish-tag <tag>', 'publish with a custom tag (instead of `latest`)')
 .action((cmd) => publish(cmd.opts()));